Add subcategories to docs for display on the Terraform Registry (#118)
We will soon be displaying documentation for this provider both on terraform.io and on the Terraform Registry.

Documentation displayed on the Terraform Registry will not use the ERB layout containing the existing navigation, and will instead build the navigation dynamically based on the directory and YAML frontmatter of a file. For providers that group similar resources and data sources by service or use case (subcategories), we'll need to add this information to the Markdown source file.

This PR modifies Resource and Data Source documentation source files by adding a subcategory to the metadata if those files were grouped previously.

For more information about how documentation is rendered on the Terraform Registry, please see this reference: Terraform Registry - Provider Documentation.